requirements.txt
numpy>=1.21.0
pandas>=1.3.0
tensorflow>=2.7.0
scipy>=1.7.0
pulp>=2.4
pyyaml>=5.4.1
kubernetes>=19.15.0